requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai>=1.0.0
GitPython
aiofiles